+++ tests/Integration/RequestIntegrationTest.php
@@ -318,7 +318,8 @@ final class RequestIntegrationTest extends TestCase
             (new ServerRequest('POST', new Uri('http://www.example.com/foo')))
                 ->withUploadedFiles([
                     'foo' => new UploadedFile('foo content', 123, UPLOAD_ERR_OK, 'foo.ext', 'application/text'),
-                ]),
+                ])
+                ->withBody($this->getStreamMock(123 + 321)),
             [
                 'url' => 'http://www.example.com/foo',
                 'method' => 'POST',
@@ -345,7 +346,8 @@ final class RequestIntegrationTest extends TestCase
                         new UploadedFile('foo content', 123, UPLOAD_ERR_OK, 'foo.ext', 'application/text'),
                         new UploadedFile('bar content', 321, UPLOAD_ERR_OK, 'bar.ext', 'application/octet-stream'),
                     ],
-                ]),
+                ])
+                ->withBody($this->getStreamMock(123 + 321)),
             [
                 'url' => 'http://www.example.com/foo',
                 'method' => 'POST',
@@ -381,7 +383,8 @@ final class RequestIntegrationTest extends TestCase
                             new UploadedFile('bar content', 321, UPLOAD_ERR_OK, 'bar.ext', 'application/octet-stream'),
                         ],
                     ],
-                ]),
+                ])
+                ->withBody($this->getStreamMock(123 + 321)),
             [
                 'url' => 'http://www.example.com/foo',
                 'method' => 'POST',
@@ -460,6 +463,25 @@ final class RequestIntegrationTest extends TestCase
                     'Content-Type' => ['application/json'],
                 ],
             ],
+        ];
+
+        yield [
+            [
+                'max_request_body_size' => 'always',
+            ],
+            (new ServerRequest('POST', new Uri('http://www.example.com/foo')))
+                ->withHeader('Content-Type', 'application/json')
+                ->withBody($this->getStreamMock(0)),
+            [
+                'url' => 'http://www.example.com/foo',
+                'method' => 'POST',
+                'headers' => [
+                    'Host' => ['www.example.com'],
+                    'Content-Type' => ['application/json'],
+                ],
+            ],
+            null,
+            null,
         ];
     }